iteration 2 of scout cdi: IBeanInstanceFactory
Reviewed-on: https://git.eclipse.org/r/43296
Tested-by: Ivan Motsch <ivan.motsch@bsiag.com>

+++ org.eclipse.scout.commons/src/main/java/org/eclipse/scout/commons/annotations/Priority.java
@@ -4,7 +4,7 @@
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
- * 
+ *
  * Contributors:
  *     BSI Business Systems Integration AG - initial API and implementation
  ******************************************************************************/
@@ -30,5 +30,5 @@ public @interface Priority {
   /**
    * @return The type's priority.
    */
-  float value();
+  double value();
 }
